Remove BaseTransaction and add a builder
